% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate_vec.R
\name{deflate_vec}
\alias{deflate_vec}
\title{Deflate Monetary Values by Brazilian Price Indices}
\usage{
deflate_vec(dates, values, ref, index = "IPCA")
}
\description{
Deflate Monetary Values by Brazilian Price Indices
}
