% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_autoarima_specs.R
\name{show_autoarima_specs}
\alias{show_autoarima_specs}
\title{Show ARIMA Specifications by AICc}
\usage{
show_autoarima_specs(
  .data,
  d,
  D,
  non_seas = 4,
  seas = 3,
  max_order = 4,
  stepwise = FALSE
)
}
\description{
Show ARIMA Specifications by AICc
}
