% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ic_autoarima.R
\name{show_ic_autoarima}
\alias{show_ic_autoarima}
\title{Show ARIMA Specifications by AICc}
\usage{
show_ic_autoarima(
  .data,
  non_seas = 4,
  seas = 2,
  max_order = 6,
  stepwise = FALSE
)
}
\description{
Show ARIMA Specifications by AICc
}
